refactor: err message
