+++ src/common/constant/error-message.ts
@@ -1,0 +1,12 @@
+export enum ErrorMessage {
+  UNAUTHORIZED = 'Unauthorized',
+  NOT_FOUND = 'Not found',
+
+  INVALID_ANSWER_TYPE = 'Invalid answer type',
+  INVALID_AB_ANSWER_TYPE = 'Answers must be A or B',
+  ALREADY_ANSWERED = 'Already answered',
+
+  INVALID_OBJECT_ID = 'Invalid objectId',
+
+  USER_API_ERR = 'User API Connection Err',
+}
