+++ src/common/pipes/parse-object-id.pipe.ts
@@ -1,3 +1,4 @@
+import { ErrorMessage } from '@/common/constant/error-message';
 import { Types } from 'mongoose';
 import { PipeTransform, Injectable, BadRequestException } from '@nestjs/common';
 
@@ -7,7 +8,7 @@ export class ParseObjectIdPipe implements PipeTransform<any, Types.ObjectId> {
     try {
       return new Types.ObjectId(value);
     } catch (error) {
-      throw new BadRequestException('Validation failed (ObjectId is expected)');
+      throw new BadRequestException(ErrorMessage.INVALID_OBJECT_ID);
     }
   }
 }
