+++ src/schema/file.schema.ts
@@ -1,4 +1,4 @@
-import { Status } from '@/common/enum';
+import { Status } from '@/common/constant/enum';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
